Update to d020a0f68d1779c154ecfc1e5bf6cae051f90e80

+++ php/elFinderVolumeDriver.class.php
@@ -227,10 +227,12 @@ abstract class elFinderVolumeDriver {
 		'accessControl'   => null,
 		// some data required by access control
 		'accessControlData' => null,
-		// default permissions. not set hidden/locked here - take no effect
+		// default permissions.
 		'defaults'     => array(
 			'read'   => true,
-			'write'  => true
+			'write'  => true,
+			'locked' => false,
+			'hidden' => false
 		),
 		// files attributes
 		'attributes'   => array(),
@@ -391,6 +393,7 @@ abstract class elFinderVolumeDriver {
 		'bz'    => 'application/x-bzip2',
 		'bz2'   => 'application/x-bzip2',
 		'tbz'   => 'application/x-bzip2',
+		'xz'    => 'application/x-xz',
 		'zip'   => 'application/zip',
 		'rar'   => 'application/x-rar',
 		'tar'   => 'application/x-tar',
@@ -611,8 +614,8 @@ abstract class elFinderVolumeDriver {
 		$this->defaults = array(
 			'read'    => isset($this->options['defaults']['read'])  ? !!$this->options['defaults']['read']  : true,
 			'write'   => isset($this->options['defaults']['write']) ? !!$this->options['defaults']['write'] : true,
-			'locked'  => false,
-			'hidden'  => false
+			'locked'  => isset($this->options['defaults']['locked']) ? !!$this->options['defaults']['locked'] : false,
+			'hidden'  => isset($this->options['defaults']['hidden']) ? !!$this->options['defaults']['hidden'] : false
 		);
 
 		// root attributes
@@ -788,7 +791,7 @@ abstract class elFinderVolumeDriver {
 			$this->tmbURL .= '/';
 		}
 		
-		$this->nameValidator = is_string($this->options['acceptedName']) && !empty($this->options['acceptedName']) 
+		$this->nameValidator = !empty($this->options['acceptedName']) && (is_string($this->options['acceptedName']) || is_callable($this->options['acceptedName']))
 			? $this->options['acceptedName']
 			: '';
 
@@ -1056,14 +1059,6 @@ abstract class elFinderVolumeDriver {
 		$path = $this->decode($hash);
 		
 		return ($file = $this->stat($path)) ? $file : $this->setError(elFinder::ERROR_FILE_NOT_FOUND);
-		
-		if (($file = $this->stat($path)) != false) {
-			if ($realpath) {
-				$file['realpath'] = $path;
-			}
-			return $file;
-		}
-		return $this->setError(elFinder::ERROR_FILE_NOT_FOUND);
 	}
 	
 	/**
@@ -1754,6 +1749,7 @@ abstract class elFinderVolumeDriver {
 		}
 		
 		$work_path = $this->getWorkFile($this->encoding? $this->convEncIn($path, true) : $path);
+
 		if (!$work_path || !is_writable($work_path)) {
 			if ($work_path && $path !== $work_path && is_file($work_path)) {
 				@unlink($work_path);
@@ -1828,7 +1824,7 @@ abstract class elFinderVolumeDriver {
 	 **/
 	public function rm($hash) {
 		return $this->commandDisabled('rm')
-			? array(elFinder::ERROR_ACCESS_DENIED)
+			? $this->setError(elFinder::ERROR_PERM_DENIED)
 			: $this->remove($this->decode($hash));
 	}
 	
@@ -1888,6 +1884,8 @@ abstract class elFinderVolumeDriver {
 			return $this->tmpPath;
 		} else if (@ $this->tmp) {
 			return $this->tmp;
+		} else if (function_exists('sys_get_temp_dir')) {
+			return sys_get_temp_dir();
 		} else if (@ $this->tmbPath) {
 			return $this->tmbPath;
 		} else {
@@ -2193,7 +2191,7 @@ abstract class elFinderVolumeDriver {
 	}
 	
 	/**
-	 * Validate file name based on $this->options['acceptedName'] regexp
+	 * Validate file name based on $this->options['acceptedName'] regexp or function
 	 *
 	 * @param  string  $name  file name
 	 * @return bool
@@ -2201,12 +2199,13 @@ abstract class elFinderVolumeDriver {
 	 **/
 	protected function nameAccepted($name) {
 		if ($this->nameValidator) {
-			if (function_exists($this->nameValidator)) {
-				$f = $this->nameValidator;
-				return $f($name);
+			if (is_callable($this->nameValidator)) {
+				$res = call_user_func($this->nameValidator, $name);
+				return $res;
 			}
-
-			return preg_match($this->nameValidator, $name);
+			if (preg_match($this->nameValidator, '') !== false) {
+				return preg_match($this->nameValidator, $name);
+			}
 		}
 		return true;
 	}
@@ -2317,6 +2316,41 @@ abstract class elFinderVolumeDriver {
 	/*********************** util mainly for inheritance class *********************/
 	
 	/**
+	 * Get temporary filename. Tempfile will be removed when after script execution finishes or exit() is called.
+	 * When needing the unique file to a path, give $path to parameter.
+	 * 
+	 * @param  string       $path for get unique file to a path
+	 * @return string|false
+	 * @author Naoki Sawada
+	 */
+	protected function getTempFile($path = '') {
+		static $cache = array();
+		static $rmfunc;
+		
+		$key = '';
+		if ($path !== '') {
+			$key = $this->id . '#' . $path;
+			if (isset($cache[$key])) {
+				return $cache[$key];
+			}
+		}
+		
+		if ($tmpdir = $this->getTempPath()) {
+			if (!$rmfunc) {
+				$rmfunc = create_function('$f', '@unlink($f);');
+			}
+			$name = tempnam($tmpdir, 'ELF');
+			if ($key) {
+				$cache[$key] = $name;
+			}
+			register_shutdown_function($rmfunc, $name);
+			return $name;
+		}
+		
+		return false;
+	}
+	
+	/**
 	 * File path of local server side work file path
 	 * 
 	 * @param  string $path path need convert encoding to server encoding
@@ -2324,15 +2358,16 @@ abstract class elFinderVolumeDriver {
 	 * @author Naoki Sawada
 	 */
 	protected function getWorkFile($path) {
-		$work = tempnam(sys_get_temp_dir(), 'elf');
-		if ($wfp = fopen($work, 'wb')) {
-			if ($fp = $this->_fopen($path)) {
-				while(!feof($fp)) {
-					fwrite($wfp, fread($fp, 8192));
+		if ($work = $this->getTempFile()) {
+			if ($wfp = fopen($work, 'wb')) {
+				if ($fp = $this->_fopen($path)) {
+					while(!feof($fp)) {
+						fwrite($wfp, fread($fp, 8192));
+					}
+					$this->_fclose($fp, $path);
+					fclose($wfp);
+					return $work;
 				}
-				$this->_fclose($fp, $path);
-				fclose($wfp);
-				return $work;
 			}
 		}
 		return false;
@@ -2342,16 +2377,19 @@ abstract class elFinderVolumeDriver {
 	 * Get image size array with `dimensions`
 	 *
 	 * @param string $path path need convert encoding to server encoding
+	 * @param string $mime file mime type
 	 * @return array|false
 	 */
-	public function getImageSize($path) {
+	public function getImageSize($path, $mime = '') {
 		$size = false;
-		if ($work = $this->getWorkFile($path)) {
-			if ($size = @getimagesize($work)) {
-				$size['dimensions'] = $size[0].'x'.$size[1];
+		if ($mime === '' || strtolower(substr($mime, 0, 5)) === 'image') {
+			if ($work = $this->getWorkFile($path)) {
+				if ($size = @getimagesize($work)) {
+					$size['dimensions'] = $size[0].'x'.$size[1];
+				}
 			}
+			is_file($work) && @unlink($work);
 		}
-		is_file($work) && @unlink($work);
 		return $size;
 	}
 	
@@ -2509,7 +2547,7 @@ abstract class elFinderVolumeDriver {
 		$stat['write'] = intval($this->attr($path, 'write', isset($stat['write']) ? !!$stat['write'] : null, $isDir));
 		if ($root) {
 			$stat['locked'] = 1;
-		} elseif ($this->attr($path, 'locked', !empty($stat['locked']), $isDir)) {
+		} elseif ($this->attr($path, 'locked', isset($stat['locked']) ? !!$stat['locked'] : null, $isDir)) {
 			$stat['locked'] = 1;
 		} else {
 			unset($stat['locked']);
@@ -2517,9 +2555,9 @@ abstract class elFinderVolumeDriver {
 
 		if ($root) {
 			unset($stat['hidden']);
-		} elseif ($this->attr($path, 'hidden', !empty($stat['hidden']), $isDir) 
+		} elseif ($this->attr($path, 'hidden', isset($stat['hidden']) ? !!$stat['hidden'] : null, $isDir) 
 		|| !$this->mimeAccepted($stat['mime'])) {
-			$stat['hidden'] = $root ? 0 : 1;
+			$stat['hidden'] = 1;
 		} else {
 			unset($stat['hidden']);
 		}
@@ -3661,7 +3699,6 @@ abstract class elFinderVolumeDriver {
 			$arcs['extract']['application/x-tar'] = array('cmd' => 'tar', 'argc' => '-xf', 'ext' => 'tar');
 			unset($o);
 			$test = $this->procExec('gzip --version', $o, $c);
-		
 			if ($c == 0) {
 				$arcs['create']['application/x-gzip']  = array('cmd' => 'tar', 'argc' => '-czf', 'ext' => 'tgz');
 				$arcs['extract']['application/x-gzip'] = array('cmd' => 'tar', 'argc' => '-xzf', 'ext' => 'tgz');
@@ -3671,6 +3708,12 @@ abstract class elFinderVolumeDriver {
 			if ($c == 0) {
 				$arcs['create']['application/x-bzip2']  = array('cmd' => 'tar', 'argc' => '-cjf', 'ext' => 'tbz');
 				$arcs['extract']['application/x-bzip2'] = array('cmd' => 'tar', 'argc' => '-xjf', 'ext' => 'tbz');
+			}
+			unset($o);
+			$test = $this->procExec('xz --version', $o, $c);
+			if ($c == 0) {
+				$arcs['create']['application/x-xz']  = array('cmd' => 'tar', 'argc' => '-cJf', 'ext' => 'xz');
+				$arcs['extract']['application/x-xz'] = array('cmd' => 'tar', 'argc' => '-xJf', 'ext' => 'xz');
 			}
 		}
 		unset($o);
@@ -3700,27 +3743,15 @@ abstract class elFinderVolumeDriver {
 		if ($c == 0) {
 			$arcs['create']['application/x-7z-compressed']  = array('cmd' => '7za', 'argc' => 'a', 'ext' => '7z');
 			$arcs['extract']['application/x-7z-compressed'] = array('cmd' => '7za', 'argc' => 'e -y', 'ext' => '7z');
-				
-			if (empty($arcs['create']['application/x-gzip'])) {
-				$arcs['create']['application/x-gzip'] = array('cmd' => '7za', 'argc' => 'a -tgzip', 'ext' => 'tar.gz');
-			}
-			if (empty($arcs['extract']['application/x-gzip'])) {
-				$arcs['extract']['application/x-gzip'] = array('cmd' => '7za', 'argc' => 'e -tgzip -y', 'ext' => 'tar.gz');
-			}
-			if (empty($arcs['create']['application/x-bzip2'])) {
-				$arcs['create']['application/x-bzip2'] = array('cmd' => '7za', 'argc' => 'a -tbzip2', 'ext' => 'tar.bz');
-			}
-			if (empty($arcs['extract']['application/x-bzip2'])) {
-				$arcs['extract']['application/x-bzip2'] = array('cmd' => '7za', 'argc' => 'a -tbzip2 -y', 'ext' => 'tar.bz');
-			}
+			
 			if (empty($arcs['create']['application/zip'])) {
-				$arcs['create']['application/zip'] = array('cmd' => '7za', 'argc' => 'a -tzip -l', 'ext' => 'zip');
+				$arcs['create']['application/zip'] = array('cmd' => '7za', 'argc' => 'a -tzip', 'ext' => 'zip');
 			}
 			if (empty($arcs['extract']['application/zip'])) {
 				$arcs['extract']['application/zip'] = array('cmd' => '7za', 'argc' => 'e -tzip -y', 'ext' => 'zip');
 			}
 			if (empty($arcs['create']['application/x-tar'])) {
-				$arcs['create']['application/x-tar'] = array('cmd' => '7za', 'argc' => 'a -ttar -l', 'ext' => 'tar');
+				$arcs['create']['application/x-tar'] = array('cmd' => '7za', 'argc' => 'a -ttar', 'ext' => 'tar');
 			}
 			if (empty($arcs['extract']['application/x-tar'])) {
 				$arcs['extract']['application/x-tar'] = array('cmd' => '7za', 'argc' => 'e -ttar -y', 'ext' => 'tar');
@@ -3730,38 +3761,20 @@ abstract class elFinderVolumeDriver {
 			unset($o);
 			$this->procExec('7z', $o, $c);
 			if ($c == 0) {
-				$arcs['create']['application/x-7z-compressed']  = array('cmd' => '7z', 'argc' => 'a -mx0', 'ext' => '7z');
-				$arcs['extract']['application/x-7z-compressed'] = array('cmd' => '7z', 'argc' => 'x -y', 'ext' => '7z');
-					
-				if (empty($arcs['create']['application/x-gzip'])) {
-					$arcs['create']['application/x-gzip'] = array('cmd' => '7z', 'argc' => 'a -tgzip -mx0', 'ext' => 'tar.gz');
-				}
-				if (empty($arcs['extract']['application/x-gzip'])) {
-					$arcs['extract']['application/x-gzip'] = array('cmd' => '7z', 'argc' => 'x -tgzip -y', 'ext' => 'tar.gz');
-				}
-				if (empty($arcs['create']['application/x-bzip2'])) {
-					$arcs['create']['application/x-bzip2'] = array('cmd' => '7z', 'argc' => 'a -tbzip2 -mx0', 'ext' => 'tar.bz');
-				}
-				if (empty($arcs['extract']['application/x-bzip2'])) {
-					$arcs['extract']['application/x-bzip2'] = array('cmd' => '7z', 'argc' => 'x -tbzip2 -y', 'ext' => 'tar.bz');
-				}
+				$arcs['create']['application/x-7z-compressed']  = array('cmd' => '7z', 'argc' => 'a', 'ext' => '7z');
+				$arcs['extract']['application/x-7z-compressed'] = array('cmd' => '7z', 'argc' => 'e -y', 'ext' => '7z');
+				
 				if (empty($arcs['create']['application/zip'])) {
-					$arcs['create']['application/zip'] = array('cmd' => '7z', 'argc' => 'a -tzip -l -mx0', 'ext' => 'zip');
+					$arcs['create']['application/zip'] = array('cmd' => '7z', 'argc' => 'a -tzip', 'ext' => 'zip');
 				}
 				if (empty($arcs['extract']['application/zip'])) {
-					$arcs['extract']['application/zip'] = array('cmd' => '7z', 'argc' => 'x -tzip -y', 'ext' => 'zip');
+					$arcs['extract']['application/zip'] = array('cmd' => '7z', 'argc' => 'e -tzip -y', 'ext' => 'zip');
 				}
 				if (empty($arcs['create']['application/x-tar'])) {
-					$arcs['create']['application/x-tar'] = array('cmd' => '7z', 'argc' => 'a -ttar -l -mx0', 'ext' => 'tar');
+					$arcs['create']['application/x-tar'] = array('cmd' => '7z', 'argc' => 'a -ttar', 'ext' => 'tar');
 				}
 				if (empty($arcs['extract']['application/x-tar'])) {
-					$arcs['extract']['application/x-tar'] = array('cmd' => '7z', 'argc' => 'x -ttar -y', 'ext' => 'tar');
-				}
-				if (empty($arcs['create']['application/x-rar'])) {
-					$arcs['create']['application/x-rar']  = array('cmd' => '7z', 'argc' => 'a -trar -l -mx0', 'ext' => 'rar');
-				}
-				if (empty($arcs['extract']['application/x-rar'])) {
-					$arcs['extract']['application/x-rar'] = array('cmd' => '7z', 'argc' => 'x -trar -y', 'ext' => 'rar');
+					$arcs['extract']['application/x-tar'] = array('cmd' => '7z', 'argc' => 'e -ttar -y', 'ext' => 'tar');
 				}
 			}
 		}
